config/requirements.txt
asgiref==3.3.1
Django==1.11.17
django-rest-auth==0.9.5
djangorestframework==3.8.2
Pillow==8.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
